% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data_emis_2km}
\alias{data_emis_2km}
\title{EMIS 2km data description}
\format{A data frame with 18 rows and 8 variables}
\source{
\url{http://emis.jrc.ec.europa.eu/webservices/2km/wcs-t}
}
\description{
A dataframe containing the description of the data available in EMIS
at 2 km spatial resolution.
The variables are as follows:
}
\details{
\itemize{
    \item{source:}{The data source}
    \item{resolution:}{Spatial resolution of the data (pixel size)}
    \item{name:}{Name of the parameters}
    \item{shortname:}{Shortname of the parameters used in the wcs-t service}
    \item{unit:}{unit}
    \item{bbox:}{Geographical bounding box (xmin,ymin,xmax,ymax)}
    \item{startdate:}{Starting date of the time series}
    \item{enddate:}{End date of the time series}
}
}

