% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{EMISR}
\alias{EMISR}
\alias{EMISR-package}
\title{EMIS and GMIS data extraction and analyses}
\description{
A package to extract and to analyse the data of the Environmental
Marine Information System (EMIS) and the Global Marine Information System
(GMIS). EMIS and GMIS are two activities of the European Commission - DG
JRC (Joint Research Centre), developed within the Water Resources Unit of
the Institute for Environment and Sustainability (IES).
The name of this package originates from a wordplay in french:
EMISR reads in french sounds like "\'emissaire" or emissary.
}

