% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getemisdataseries}
\alias{getemisdataseries}
\title{This function downloads a time series of a variable from EMIS for a period of time month on a spatial area
using the wcs-t service}
\usage{
getemisdataseries(name = "EMIS_A_CHLA", resolution = "4km",
  startdate = "2005-09", enddate = "2005-10", xmin = 15, xmax = 20.5,
  ymin = 30, ymax = 32.5)
}
\arguments{
\item{name}{A character vector of the shortname of the variable}

\item{resolution}{A character vector giving the spatial resolution of the data: "4km" or "2km"}

\item{startdate}{A character vector of the month who begins the series("YYYY-MM")}

\item{enddate}{A character vector of the month who ends the series("YYYY-MM")}

\item{xmin}{A numeric vector of the lower longitude of the spatial area}

\item{xmax}{A numeric vector of the max longitude of the spatial area}

\item{ymin}{A numeric vector of the lower latitude of the spatial area}

\item{ymax}{A numeric vector of the max latitude of the spatial area}
}
\value{
A rasterstack object of the variable for the given period
}
\description{
This function downloads a time series of a variable from EMIS for a period of time month on a spatial area
using the wcs-t service
}
\examples{
\dontrun{
	#extraction of the MODIS chlorophyll a concentration monthly maps for 2005 and 2006 on the gulf of Syrte (Lybia)
	img<-getemisdataseries("EMIS_A_CHLA","4km","2005-01","2006-12",15,20.5,30,32.5)
	plot(img)
  }
}
\keyword{EMIS}
\keyword{wcs-t}

