% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getgmisdata}
\alias{getgmisdata}
\title{This function downloads a variable from GMIS for a month on a spatial area
using the wcs-t service}
\usage{
getgmisdata(name = "GMIS_A_CHLA", resolution = "4km", date = "2005-09",
  xmin = 15, xmax = 20.5, ymin = 30, ymax = 32.5)
}
\arguments{
\item{name}{A character vector of the shortname of the variable}

\item{resolution}{A character vector giving the spatial resolution of the data: "4km" or "9km"}

\item{date}{A character vector of the month ("YYYY-MM")}

\item{xmin}{A numeric vector of the lower longitude of the spatial area}

\item{xmax}{A numeric vector of the max longitude of the spatial area}

\item{ymin}{A numeric vector of the lower latitude of the spatial area}

\item{ymax}{A numeric vector of the max latitude of the spatial area}
}
\value{
A rasterlayer object of the map of the variable for the given month
}
\description{
This function downloads a variable from GMIS for a month on a spatial area
using the wcs-t service
}
\examples{
\dontrun{
	#extraction of the MODIS chlorophyll a concentration for Septembre 2009 in the vicinity of the Bird Island area (South Africa)
	img<-getgmisdata("GMIS_A_CHLA","4km","2005-09",xmin=26.1,xmax=26.5,ymin=-34,ymax=-33.5)
	plot(img)
  }
}
\keyword{GMIS}
\keyword{wcs-t}

