% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getgmisdataseries}
\alias{getgmisdataseries}
\title{This function downloads a time series of a variable from GMIS for a period of time on a spatial area
using the wcs-t service}
\usage{
getgmisdataseries(name = "GMIS_A_CHLA", resolution = "4km",
  startdate = "2005-09", enddate = "2005-10", xmin = 15, xmax = 20.5,
  ymin = 30, ymax = 32.5)
}
\arguments{
\item{name}{A character vector of the shortname of the variable}

\item{resolution}{A character vector giving the spatial resolution of the data: "4km" or "9km"}

\item{startdate}{A character vector of the month who begins the series("YYYY-MM")}

\item{enddate}{A character vector of the month who ends the series("YYYY-MM")}

\item{xmin}{A numeric vector of the lower longitude of the spatial area}

\item{xmax}{A numeric vector of the max longitude of the spatial area}

\item{ymin}{A numeric vector of the lower latitude of the spatial area}

\item{ymax}{A numeric vector of the max latitude of the spatial area}
}
\value{
A rasterstack object of the variable for the given period
}
\description{
This function downloads a time series of a variable from GMIS for a period of time on a spatial area
using the wcs-t service
}
\examples{
\dontrun{
     #extraction of the MODIS chlorophyll a concentration for 2008 and 2010 in the vicinity of the Bird Island area (South Africa)
     img<-getgmisdataseries("GMIS_A_CHLA","4km","2008-01","2010-12",xmin=26.1,xmax=26.5,ymin=-34,ymax=-33.5)
     plot(img)
  }
}
\keyword{GMIS}
\keyword{wcs-t}

