% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mpa2}
\alias{mpa2}
\title{Marine protected area database description}
\format{A SpatialPolygonsDataFrame with 5612 features (the last 5612 MPA according to their wdpaid)}
\source{
\url{http://www.protectedplanet.net}
}
\description{
The shapefile of the marine protected area (MPA) was downloaded from
\url{http://www.protectedplanet.net} the 03/12/2014.
The shapefile was converted in SpatialPolygonsDataFrame using rgdal:
\code{mpa<-readOGR("EMISR/data/mpa/","mpa",encoding="Latin1")}
then mpa was splitted in two: mpa1 and mpa2.
}
\details{
No projection system is available for this SpatialPolygonsDataFrame
due to coordinates problem.
26 variables are available, describing the mpa characteristics (see reference).
}
\references{
UNEP-WCMC (2012). Data Standards for the World Database on Protected Areas. UNEP-WCMC: Cambridge, UK.
}

