% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mpaextract}
\alias{mpaextract}
\title{This function downloads a time series of a variable from EMIS for a period of time month for a given MPA
using the wcs-t service. The data are extracted with a buffer of about 10 km around the MPA.}
\usage{
mpaextract(name = "EMIS_A_CHLA", resolution = "4km",
  startdate = "2005-09", enddate = "2005-10", wdpaid = 555540558)
}
\arguments{
\item{name}{A character vector of the shortname of the variable}

\item{resolution}{A character vector giving the spatial resolution of the data: "4km" or "2km"}

\item{startdate}{A character vector of the month who begins the series("YYYY-MM")}

\item{enddate}{A character vector of the month who ends the series("YYYY-MM")}

\item{wdpaid}{An integer of the ID of the marine protected area corresponding to the wdpaid of the mpa}
}
\value{
A rasterstack object of the variable for the given period on the MPA
}
\description{
This function downloads a time series of a variable from EMIS for a period of time month for a given MPA
using the wcs-t service. The data are extracted with a buffer of about 10 km around the MPA.
}
\examples{
\dontrun{
     #extraction of the MODIS sea surface temperature at 2 km in 2009 and 2012 on the Pantelleria mpa (Italy)
     img<-mpaextract("EMIS_T_SST","2km","2009-01","2012-12",555540558)
     plot(img)
  }
}
\keyword{EMIS}
\keyword{wcs-t}

