% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mpaprocess}
\alias{mpaprocess}
\title{This function compute a set of indicators based on a series of map of a EMIS/GMIS parameter
and draw some maps and plots related to the parameter.}
\usage{
mpaprocess(name = "EMIS_A_CHLA", resolution = "4km",
  startdate = "2005-09", enddate = "2005-10", wdpaid = 555540558)
}
\arguments{
\item{name}{A character vector of the shortname of the variable}

\item{resolution}{A character vector giving the spatial resolution of the data: "4km" or "2km"}

\item{startdate}{A character vector of the month who begins the series("YYYY-MM")}

\item{enddate}{A character vector of the month who ends the series("YYYY-MM")}

\item{wdpaid}{An integer of the ID of the marine protected area corresponding to the wdpaid of the mpa}
}
\value{
A list of object containing the plot and the statistics
}
\description{
The parameters are monthly data available on EMIS or GMIS.
"name" gives the short name of the parameter processed.
"timerange" is the time range of the parameter.
"pixelnb" gives the number of 4 by 4 km pixels covering the study area: this area is the geographical extent of the MPA polygon (ie a rectangle encompassing the MPA).
"fracnbna" is the percentage of missing data in the MPA polygon over the time range.
"fractsna" is the percentage of missing value for the time series of the parameters. This time serie is calculated using the monthly mean of the pixel values INSIDE the MPA polygon.
"mean" is the average value of the parameter INSIDE the MPA polygon over the time range.
"sd" is the standard deviation of the parameter values INSIDE the MPA polygon over the time range.
"min" is the minimum of the parameter values INSIDE the MPA polygon over the time range.
"max" is the maximum of the parameter values INSIDE the MPA polygon over the time range.
The monthly time series X(t) of the parameter is decomposed as  X(t)=S(t)+T(t)+I(t) (see Vantrepotte and Melin, 2009), where S, T, and I represent, respectively, a seasonal signal, a trend, and an irregular (or residual) component (or remainder). These components are presented in the figure seasonal composition. In this framework:
"varseason" is the percentage of variance associated the seasonal component and "vartrend" the percentage of variance associated with the linear component (the trend).
A statistical test (Kendall's tau test) is performed on the linear component in order to assess the significance of a monotic trend:
"trendtest" gives the p.value associated to this test.
"senslope" gives the value of the Sen slope (alternately, Theil slope). This value is the median
slope joining all pairs of observations and is expressed by
quantity per year.
References:
Vantrepotte, V. & Melin, F. Temporal variability of 10-year global SeaWiFS time-series of phytoplankton chlorophyll a concentration ICES J. Mar. Sci., 2009, 66, 1547-1556
}
\examples{
\dontrun{
     #analysis of the MODIS sea surface temperature at 2 km 2005 and 2006 on the Pantelleria mpa (Italy)
	pltstat<-mpaprocess("EMIS_T_SST","2km","2005-01","2007-12",555540558)
	#map of the whole series
	pltstat[[1]]$pltall
	#map of the average SST
	pltstat[[1]]$pltmean
	#map of the climatology
	pltstat[[1]]$pltclim
	#some statistics
	pltstat[[2]]$stat
	#time series decomposition of the parameters averaged on the MPA area
	pltstat[[2]]$pltts
	#extraction of the MODIS sea surface temperature at 4 km from GMIS for 2005-2007 in the the Bird Island area marine protected area (South Africa)
	pltstat<-mpaprocess("GMIS_T_SST","4km","2005-01","2007-12",306180)
	#map of the whole series
	pltstat[[1]]$pltall
	#map of the average SST
	pltstat[[1]]$pltmean
	#map of the climatology
	pltstat[[1]]$pltclim
	#some statistics
	pltstat[[2]]$stat
	#time series decomposition of the parameters averaged on the MPA area
	pltstat[[2]]$pltts
  }
}
\keyword{EMIS}
\keyword{wcs-t}

