% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mpaprocessplot}
\alias{mpaprocessplot}
\title{This function draw figures based on a series of map of a EMIS/GMIS parameter for a marine protected area}
\usage{
mpaprocessplot(imgs, mpa, name, unite, logscale)
}
\arguments{
\item{imgs}{A rasterstack of the series of map extracted using mpaextract}

\item{mpa}{A SpatialPolygonsDataFrame of the marine protected area of interest}

\item{name}{A character vector of the variable name}

\item{unite}{A character vector of the unit of variable name}

\item{logscale}{A boolean defining the use of a logscale on a graph (TRUE) or not (FALSE)}
}
\value{
A list of plots
}
\description{
This function draw figures based on a series of map of a EMIS/GMIS parameter for a marine protected area
}
\examples{
\dontrun{
     #analysis of the MODIS sea surface temperature at 2 km 2009 and 2012 on the Pantelleira mpa (Italy)
	plt<-mpaprocessplot(imgs=pantelleria_sst,mpa=pantelleria_mpa,name="EMIS_T_SST",unite="oC",logscale=FALSE)
	#map of the whole series
	plt[[1]]
	#map of the average SST
	plt[[2]]
	#map of the climatology
	plt[[3]]
	#boxplot of the climatology
	plt[[4]]
  }
}
\keyword{EMIS}
\keyword{wcs-t}

