% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mpaprocessstat}
\alias{mpaprocessstat}
\title{This function compute some indicators for a series of map of a EMIS/GMIS parameter for a marine protected area}
\usage{
mpaprocessstat(imgs, mpa, name, unite)
}
\arguments{
\item{imgs}{A rasterstack of the series of map extracted using mpaextract}

\item{mpa}{A SpatialPolygonsDataFrame of the marine protected area of interest}

\item{name}{A character vector of the variable name}

\item{unite}{A character vector of the unit of variable name}
}
\value{
A list of 2 objects: a dataframe and a plot
}
\description{
The parameters are monthly data available on EMIS or GMIS.
"name" gives the short name of the parameter processed.
"timerange" is the time range of the parameter.
"pixelnb" gives the number of 4 by 4 km pixels covering the study area: this area is the geographical extent of the MPA polygon (ie a rectangle encompassing the MPA).
"fracnbna" is the percentage of missing data in the MPA polygon over the time range.
"fractsna" is the percentage of missing value for the time series of the parameters. This time serie is calculated using the monthly mean of the pixel values INSIDE the MPA polygon.
"mean" is the average value of the parameter INSIDE the MPA polygon over the time range.
"sd" is the standard deviation of the parameter values INSIDE the MPA polygon over the time range.
"min" is the minimum of the parameter values INSIDE the MPA polygon over the time range.
"max" is the maximum of the parameter values INSIDE the MPA polygon over the time range.
The monthly time series X(t) of the parameter is decomposed as  X(t)=S(t)+T(t)+I(t) (see Vantrepotte and Melin, 2009), where S, T, and I represent, respectively, a seasonal signal, a trend, and an irregular (or residual) component (or remainder). These components are presented in the figure seasonal composition. In this framework:
"varseason" is the percentage of variance associated the seasonal component and "vartrend" the percentage of variance associated with the linear component (the trend).
A statistical test (Kendall's tau test) is performed on the linear component in order to assess the significance of a monotic trend:
"trendtest" gives the p.value associated to this test.
"senslope" gives the value of the Sen slope (alternately, Theil slope). This value is the median
slope joining all pairs of observations and is expressed by
quantity per year.
References:
Vantrepotte, V. & Melin, F. Temporal variability of 10-year global SeaWiFS time-series of phytoplankton chlorophyll a concentration ICES J. Mar. Sci., 2009, 66, 1547-1556
}
\examples{
#analysis of the MODIS sea surface temperature at 2 km 2009 and 2012 on the Pantelleira mpa (Italy)
	stat<-mpaprocessstat(imgs=pantelleria_sst,mpa=pantelleria_mpa,name="EMIS_T_SST",unite="oC")
	#statistics
	stat[[1]]
	#plot of the time series decomposition of the parameter averaged on the MPA surface
	stat[[2]]
}
\keyword{EMIS}
\keyword{wcs-t}

