% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pantelleria_chl}
\alias{pantelleria_chl}
\title{Chlorophyll a surface concentration of the Pantelleria Marine Protected Area}
\format{A RasterStack with 48 layers}
\source{
\url{http://emis.jrc.ec.europa.eu/emis_6_0.php}
}
\description{
The chl was downloaded using:
\code{pantelleria_chl<-mpaextract("EMIS_A_CHLA","2km","2009-01","2012-12",555540558)}
}

