% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pantelleria_mpa}
\alias{pantelleria_mpa}
\title{Pantelleria Marine Protecte Area}
\format{A SpatialPolygonsDataFrame}
\source{
\url{http://www.protectedplanet.net}
}
\description{
The polygon was extracted from the mpa2 file using the wdpid 555540558
}

