% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pantelleria_sst}
\alias{pantelleria_sst}
\title{Sea surface temperature of the Pantelleria Marine Protected Area}
\format{A RasterStack with 48 layers}
\source{
\url{http://emis.jrc.ec.europa.eu/emis_6_0.php}
}
\description{
The SST was downloaded using:
\code{pantelleria_sst<-mpaextract("EMIS_T_SST","2km","2009-01","2012-12",555540558)}
}

